refactor: pull out shared header styling into StyledHeader component
