+++ src/components/Header/index.ts
@@ -1,0 +1,2 @@
+export { default } from "./FixedHeader";
+export { StyledHeader } from "./StyledHeader";
